+++ src/glegacy/cxgo_GAME3_3.go
@@ -4199,7 +4199,7 @@ func nox_xxx_collideChest_4E9C40(a1 *uint32, a2 int32) {
 		return
 	}
 LABEL_14:
-	v6 = asFunc(*(*uint32)(unsafe.Add(unsafe.Pointer(a1), 4*181)), (*func(*uint32))(nil)).(func(*uint32))
+	v6 = asFunc(*(*int32)(unsafe.Add(unsafe.Pointer(a1), 4*181)), (*func(*uint32))(nil)).(func(*uint32))
 	if v6 != nil {
 		v6(a1)
 	}
@@ -5678,7 +5678,7 @@ func nox_xxx_collideAnkhQuest_4EBF40(a1 int32, a2 int32) {
 		if *(*uint32)(unsafe.Pointer(uintptr(v4 + 320))) < uint32(nox_float2int(v17)) {
 			v12 = (*uint32)(unsafe.Pointer(nox_xxx_newObjectByTypeID_4E3810(internCStr("AnkhTradable"))))
 			if v12 != nil {
-				(asFunc(*(*uint32)(unsafe.Add(unsafe.Pointer(v12), 4*177)), (*func(int32, *uint32, int32, uint32))(nil)).(func(int32, *uint32, int32, uint32)))(v2, v12, 1, 0)
+				(asFunc(*(*int32)(unsafe.Add(unsafe.Pointer(v12), 4*177)), (*func(int32, *uint32, int32, uint32))(nil)).(func(int32, *uint32, int32, uint32)))(v2, v12, 1, 0)
 			}
 			*(*uint32)(unsafe.Pointer(uintptr(a1 + 136))) = gameFrame()
 			nox_xxx_aud_501960(1004, (*nox_object_t)(unsafe.Pointer(uintptr(a1))), 0, 0)
@@ -6585,7 +6585,7 @@ func nox_xxx_drop_4ED790(a1p *nox_object_t, a2p *nox_object_t, a3 *float2) int32
 		*(*uint32)(unsafe.Add(unsafe.Pointer(a2), 4*4)) |= 0x40
 		nox_xxx_unit_511810((*nox_object_t)(unsafe.Pointer(uintptr(int32(uintptr(unsafe.Pointer(a2)))))))
 	}
-	v4 = asFunc(*(*uint32)(unsafe.Add(unsafe.Pointer(a2), 4*178)), (*func(int32, *uint32, *float2) int32)(nil)).(func(int32, *uint32, *float2) int32)
+	v4 = asFunc(*(*int32)(unsafe.Add(unsafe.Pointer(a2), 4*178)), (*func(int32, *uint32, *float2) int32)(nil)).(func(int32, *uint32, *float2) int32)
 	if v4 != nil {
 		result = v4(a1, a2, a3)
 	} else {
@@ -7833,7 +7833,7 @@ func nox_xxx_playerRespawnItem_4EF750(a1p *nox_object_t, a2 *byte, a3 *int32, a4
 	v5 = (*uint32)(unsafe.Pointer(nox_xxx_newObjectByTypeID_4E3810(a2)))
 	v6 = v5
 	if v5 != nil {
-		v7 = asFunc(*(*uint32)(unsafe.Add(unsafe.Pointer(v5), 4*172)), (*func(*uint32, uint32))(nil)).(func(*uint32, uint32))
+		v7 = asFunc(*(*int32)(unsafe.Add(unsafe.Pointer(v5), 4*172)), (*func(*uint32, uint32))(nil)).(func(*uint32, uint32))
 		if v7 != nil {
 			v7(v6, 0)
 		}
@@ -11142,7 +11142,7 @@ func nox_xxx_xfer_4F3E30(a1 uint16, a2p *nox_object_t, a3 int32) int32 {
 		nox_xxx_fileCryptReadCrcMB_426C20((*uint8)(unsafe.Pointer(&v12)), 4)
 		v5 = (*uint32)(unsafe.Pointer(nox_xxx_newObjectWithTypeInd_4E3450(int32(v4))))
 		v6 = v5
-		if v5 == nil || (asFunc(*(*uint32)(unsafe.Add(unsafe.Pointer(v5), 4*176)), (*func(*uint32, uint32) int32)(nil)).(func(*uint32, uint32) int32))(v5, 0) == 0 {
+		if v5 == nil || (asFunc(*(*int32)(unsafe.Add(unsafe.Pointer(v5), 4*176)), (*func(*uint32, uint32) int32)(nil)).(func(*uint32, uint32) int32))(v5, 0) == 0 {
 			break
 		}
 		v7 = int32(*(*uint32)(unsafe.Pointer(uintptr(a2 + 504))))
@@ -11295,7 +11295,7 @@ func nox_xxx_readObjectOldVer_4F4170(a1 int32, a2 int32, a3 int32) int32 {
 		nox_xxx_fileReadWrite_426AC0_file3_fread_impl((*uint8)(unsafe.Pointer(&v19)), 1)
 		if nox_crypt_IsReadOnly() == 1 {
 			if int32(uint8(int8(v19))) != 0 {
-				result = int32(uintptr(alloc.Calloc1(1, int(int32(uint8(int8(v19)))+1))))
+				result = int32(uintptr(alloc.Calloc1(1, uintptr(int32(uint8(int8(v19)))+1))))
 				*v3 = (*uint8)(unsafe.Pointer(uintptr(result)))
 				if result == 0 {
 					return result
@@ -11475,7 +11475,7 @@ func nox_xxx_mapReadWriteObjData_4F4530(a1p *nox_object_t, a2 int32) int32 {
 	}
 	nox_xxx_fileReadWrite_426AC0_file3_fread_impl((*uint8)(unsafe.Pointer(&a2)), 1)
 	if nox_crypt_IsReadOnly() != 1 || int32(uint8(int8(a2))) == 0 || (func() bool {
-		result = int32(uintptr(alloc.Calloc1(1, int(int32(uint8(int8(a2)))+1))))
+		result = int32(uintptr(alloc.Calloc1(1, uintptr(int32(uint8(int8(a2)))+1))))
 		return (func() int32 {
 			p := v2
 			*v2 = result
