+++ src/glegacy/groups.go
@@ -12,8 +12,8 @@ func nox_server_getFirstMapGroup_57C080() unsafe.Pointer {
 }
 
 // nox_server_scriptGetGroup_57C0A0
-func nox_server_scriptGetGroup_57C0A0(a1 int) unsafe.Pointer {
-	return GetServer().S().MapGroups.GroupByInd(a1).C()
+func nox_server_scriptGetGroup_57C0A0(a1 int32) unsafe.Pointer {
+	return GetServer().S().MapGroups.GroupByInd(int(a1)).C()
 }
 
 // sub_504600
@@ -22,12 +22,12 @@ func sub_504600(name *byte, ind uint32, typ uint8) {
 }
 
 // sub_57C130
-func sub_57C130(d *uint32, ind uint32) int32 {
+func sub_57C130(d *uint32, ind int32) int32 {
 	var arg []uint32
 	if d != nil {
 		arg = unsafe.Slice(d, 2)
 	}
-	return GetServer().S().MapGroups.Sub57C130(arg, ind)
+	return GetServer().S().MapGroups.Sub57C130(arg, uint32(ind))
 }
 
 // sub_5046A0
@@ -40,8 +40,8 @@ func sub_5046A0(d *uint32, ind uint32) int {
 }
 
 // sub_4CFFE0
-func sub_4CFFE0(sid int) *nox_object_t {
-	return asObjectC(GetServer().PendingObjByScriptID4CFFE0(sid))
+func sub_4CFFE0(sid int32) *nox_object_t {
+	return asObjectC(GetServer().PendingObjByScriptID4CFFE0(int(sid)))
 }
 
 // sub_579C60
