+++ src/glegacy/gui_main_menu.go
@@ -27,22 +27,24 @@ func sub_44E320() { Sub_44E320() }
 func winMainMenuAnimOutDoneFnc() int { return WinMainMenuAnimOutDoneFnc() }
 
 // sub_4A24C0
-func sub_4A24C0(a1 int) int { return Sub_4A24C0(a1) }
+func sub_4A24C0(a1 int) int32 { return int32(Sub_4A24C0(a1)) }
 
 // nox_game_showMainMenu_4A1C00
 func nox_game_showMainMenu_4A1C00() int { return Nox_game_showMainMenu_4A1C00() }
 
 // sub_43BE40
-func sub_43BE40(a1 int) { Sub_43BE40(a1) }
+func sub_43BE40(a1 int32) { Sub_43BE40(int(a1)) }
 
 // sub_43BE30
-func sub_43BE30() int { return Sub_43BE30() }
+func sub_43BE30() int32 { return int32(Sub_43BE30()) }
 
 // sub_4A18E0
-func sub_4A18E0(a1 unsafe.Pointer, a2, a3, a4 int) int { return Sub_4A18E0(AsWindowP(a1), a2, a3, a4) }
+func sub_4A18E0(a1 unsafe.Pointer, a2, a3, a4 int32) int32 {
+	return int32(Sub_4A18E0(AsWindowP(a1), int(a2), int(a3), int(a4)))
+}
 
 // nox_client_drawGeneralCallback_4A2200
-func nox_client_drawGeneralCallback_4A2200() int {
+func nox_client_drawGeneralCallback_4A2200() int32 {
 	if err := GetClient().DrawGeneral(false); err != nil {
 		guiLog.Println(err)
 		return 0
